Wrote readme.md, admin schema changed, fixed api errors, fixed server action errors, made edit page of careers, made websocket.js, api.js and comment.js for using guide.

+++ app/api/task/[id]/route.ts
@@ -1,5 +1,4 @@
 import Task from "@/model/Task"
-import SocketService from "@/service/socketService";
 import { NextRequest, NextResponse } from "next/server";
 import { Eligibility } from "@/utils/mongo/eligibility";
 
@@ -15,25 +14,20 @@ interface TaskType {
 export async function DELETE(req: NextRequest) {
     try {
         const cookie = req.cookies.get('admin-log');
-        if (!cookie) return NextResponse.json({ error: 'Please log in' }, { status: 401 });
+        if (!cookie) return NextResponse.json({ error: 'Please log in as admin, editor or author' }, { status: 401 });
 
         const coll = await Eligibility(cookie.value)
 
-        if (coll.role === '') return NextResponse.json({ error: 'Please log in as admin' }, { status: 401 });
+        if (coll.role !== 'Admin' && coll.role !== 'Editor' && coll.role !== 'Author') return NextResponse.json({ error: 'Please log in as admin, editor or author' }, { status: 401 });
 
         const id = req.nextUrl.pathname.slice(10, req.nextUrl.pathname.length)
 
         await Task.findByIdAndDelete(id)
 
-        const tasks = await Task.find() as TaskType[];
-
-        /*const socketService = SocketService.getInstance();
-        socketService.emit('deleteTask', { tasks: JSON.parse(JSON.stringify(tasks)) })*/
-
-
         return NextResponse.json({ success: 'Success' }, { status: 200 });
     }
     catch (err) {
+        console.log(err)
         return NextResponse.json({ error: 'Server error' }, { status: 500 });
     }
 }
@@ -41,11 +35,11 @@ export async function DELETE(req: NextRequest) {
 export async function PUT(req: NextRequest) {
     try {
         const cookie = req.cookies.get('admin-log');
-        if (!cookie) return NextResponse.json({ error: 'Please log in' }, { status: 401 });
+        if (!cookie) return NextResponse.json({ error: 'Please log in as admin, editor or author' }, { status: 401 });
 
         const coll = await Eligibility(cookie.value)
 
-        if (coll.role === '') return NextResponse.json({ error: 'Please log in as admin' }, { status: 401 });
+        if (coll.role !== 'Admin' && coll.role !== 'Editor' && coll.role !== 'Author') return NextResponse.json({ error: 'Please log in as admin, editor or author' }, { status: 401 });
 
 
         const id = req.nextUrl.pathname.slice(10, req.nextUrl.pathname.length)
@@ -55,20 +49,16 @@ export async function PUT(req: NextRequest) {
 
         if (!task.name) {
             const NewTask = await Task.findByIdAndUpdate(id, { name: coll.name })
-            if (!NewTask) return NextResponse.json({ error: 'Task deleted' }, { status: 500 })
+            if (!NewTask) return NextResponse.json({ error: 'The task does not exist' }, { status: 404 })
         }
         else {
             if (task.name === coll.name) await Task.findByIdAndUpdate(id, { name: null })
         }
 
-        const tasks = await Task.find() as TaskType[];
-
-        /*const socketService = SocketService.getInstance();
-        socketService.emit('setNameForTask', { tasks: JSON.parse(JSON.stringify(tasks)) })*/
-
         return NextResponse.json({ success: 'Success' }, { status: 200 })
     }
     catch (err) {
+        console.log(err)
         return NextResponse.json({ error: 'Server error' }, { status: 500 })
     }
 }
